@@ -1,42 +1,25 @@
-import React from "react";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Redirect,
-} from "react-router-dom";
-import "./App.css";
+import logo from './logo.svg';
+import './App.css';
 
-import { connect } from "react-redux";
-
-import Navbar from "./components/Navbar/Navbar";
-import Products from "./components/Products/Products";
-import Cart from "./components/Cart/Cart";
-import SingleItem from "./components/SingleItem/SingleItem";
-
-function App({ current }) {
+function App() {
   return (
-    <Router>
-      <div className="app">
-        <Navbar />
-        <Switch>
-          <Route exact path="/" component={Products} />
-          <Route exact path="/cart" component={Cart} />
-          {!current ? (
-            <Redirect to="/" />
-          ) : (
-            <Route exact path="/product/:id" component={SingleItem} />
-          )}
-        </Switch>
-      </div>
-    </Router>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
   );
 }
 
-const mapStateToProps = (state) => {
-  return {
-    current: state.shop.currentItem,
-  };
-};
-
-export default connect(mapStateToProps)(App);
+export default App;
